Secure secret token
It was out in the open before, which isn't too bad since there are no
cookies used by the website currently, but it has now been secured in
case cookies need to be implemented in the future. A SECRET_TOKEN
environment variable can now be set on the server. If none is found, it
will fall back to a development secret key that is available to
everyone.