Add ability to deploy instant alerts
Setting the ALERT environment variable will cause an alert to be
displayed at the top of all pages. It will default to the alert-info
class, but the ALERT_TYPE environment variable can change this- setting
it to "error", for example, will make it of the alert-error class. The
title, by default, is "Notice", but this can be altered via the
ALERT_TITLE environment variable.